refactor(cloud): split the object store registry into a vortex-cloud crate
Moves the URL -> ObjectStore registry out of vortex-io and folds
vortex-object-store-opendal into the same crate, so all cloud object
store integration lives in one place.

- vortex-cloud owns the registry (`registry` feature) and the
  OpenDAL-backed services, now an `opendal` module behind per-service
  `cos` / `oss` features.
- vortex-object-store-opendal is deleted; its files move with history.
- vortex-io drops its `object_store_registry` and `opendal` features and
  its `url` dependency, keeping only the generic object_store adapters.
- The facade exposes `vortex::cloud`, and vortex-python, vortex-jni and
  vortex-duckdb resolve through it.

The registry no longer depends on vortex-io at all. Inside the `opendal`
module the external crate is spelled `::opendal` to disambiguate it from
the module of the same name.
