+++ vortex-duckdb/src/multi_file.rs
@@ -7,6 +7,7 @@ use std::sync::LazyLock;
 use itertools::Itertools;
 use object_store::registry::ObjectStoreRegistry;
 use url::Url;
+use vortex::cloud::Registry;
 use vortex::error::VortexResult;
 use vortex::error::vortex_bail;
 use vortex::error::vortex_err;
@@ -15,7 +16,6 @@ use vortex::file::multi::parse_uri_or_path;
 use vortex::io::compat::Compat;
 use vortex::io::filesystem::FileSystemRef;
 use vortex::io::object_store::ObjectStoreFileSystem;
-use vortex::io::object_store::Registry;
 use vortex::io::runtime::BlockingRuntime;
 use vortex::layout::scan::multi::MultiLayoutDataSource;
 use vortex_utils::aliases::hash_map::HashMap;
