+++ vortex-io/src/object_store/mod.rs
@@ -3,12 +3,8 @@
 
 mod filesystem;
 mod read_at;
-#[cfg(feature = "object_store_registry")]
-mod registry;
 mod write;
 
 pub use filesystem::*;
 pub use read_at::*;
-#[cfg(feature = "object_store_registry")]
-pub use registry::*;
 pub use write::*;
