--- vortex-object-store-opendal/Cargo.toml
@@ -1,33 +1,0 @@
-[package]
-name = "vortex-object-store-opendal"
-description = "OpenDAL-backed object stores (Tencent COS, Alibaba OSS) for Vortex"
-version = { workspace = true }
-homepage = { workspace = true }
-repository = { workspace = true }
-authors = { workspace = true }
-license = { workspace = true }
-keywords = { workspace = true }
-include = { workspace = true }
-edition = { workspace = true }
-readme = { workspace = true }
-rust-version = { workspace = true }
-categories = { workspace = true }
-
-[package.metadata.docs.rs]
-all-features = true
-
-[dependencies]
-object_store = { workspace = true, features = ["cloud"] }
-object_store_opendal = { workspace = true }
-opendal = { workspace = true }
-tracing = { workspace = true }
-url = { workspace = true }
-vortex-utils = { workspace = true }
-
-[features]
-default = ["cos", "oss"]
-cos = ["opendal/services-cos"]
-oss = ["opendal/services-oss"]
-
-[lints]
-workspace = true
