+++ vortex-python/Cargo.toml
@@ -30,7 +30,7 @@ extension-module = []
 tui = ["dep:tokio", "dep:vortex-tui"]
 # OpenDAL-backed object stores (Tencent COS, Alibaba OSS) for `cos://` and `oss://` URLs. Pulls in
 # `opendal`, so it is opt-in.
-opendal = ["dep:vortex-object-store-opendal", "vortex/opendal"]
+opendal = ["dep:vortex-cloud", "vortex-cloud/opendal", "vortex/opendal"]
 
 [dependencies]
 arrow-array = { workspace = true }
@@ -59,10 +59,7 @@ vortex = { workspace = true, features = [
     "object_store_registry",
 ] }
 vortex-arrow = { workspace = true }
-vortex-object-store-opendal = { workspace = true, optional = true, features = [
-    "cos",
-    "oss",
-] }
+vortex-cloud = { workspace = true, optional = true }
 vortex-python-abi = { path = "../vortex-python-abi" }
 vortex-tui = { workspace = true, optional = true }
