+++ vortex/Cargo.toml
@@ -37,6 +37,7 @@ vortex-fastlanes = { workspace = true }
 vortex-file = { workspace = true, optional = true, default-features = true }
 vortex-flatbuffers = { workspace = true }
 vortex-fsst = { workspace = true }
+vortex-cloud = { workspace = true, optional = true }
 vortex-io = { workspace = true }
 vortex-ipc = { workspace = true }
 vortex-layout = { workspace = true }
@@ -74,9 +75,9 @@ files = ["dep:vortex-file"]
 memmap2 = ["vortex-buffer/memmap2"]
 object_store = ["vortex-file?/object_store", "vortex-io/object_store"]
 # The URL -> ObjectStore registry, plus the cloud backends it resolves URLs to.
-object_store_registry = ["object_store", "vortex-io/object_store_registry"]
+object_store_registry = ["object_store", "dep:vortex-cloud", "vortex-cloud/registry"]
 # OpenDAL-backed schemes (`cos://`, `oss://`) in the object store registry.
-opendal = ["object_store_registry", "vortex-io/opendal"]
+opendal = ["object_store_registry", "vortex-cloud/opendal"]
 tokio = [
     "vortex-error/tokio",
     "vortex-file?/tokio",
